Misc spacing fixes in deeplearning4j-nn Remove old helpers Update long indices

+++ deeplearning4j/deeplearning4j-nn/src/main/java/org/deeplearning4j/nn/conf/layers/GlobalPoolingLayer.java
@@ -46,7 +46,7 @@ import java.util.Map;
 public class GlobalPoolingLayer extends NoParamLayer {
 
     private PoolingType poolingType;
-    private int[] poolingDimensions;
+    private long[] poolingDimensions;
     private int pnorm;
     private boolean collapseDimensions = true;
 
@@ -131,12 +131,12 @@ public class GlobalPoolingLayer extends NoParamLayer {
 
     @Override
     public void setNIn(InputType inputType, boolean override) {
-        if(inputType.getType() == InputType.Type.CNN){
+        if(inputType.getType() == InputType.Type.CNN) {
             InputType.InputTypeConvolutional c = (InputType.InputTypeConvolutional) inputType;
             if(c.getFormat() == CNN2DFormat.NCHW){
-                poolingDimensions = new int[]{2,3};
+                poolingDimensions = new long[]{2,3};
             } else {
-                poolingDimensions = new int[]{1,2};
+                poolingDimensions = new long[]{1,2};
             }
         }
     }
@@ -220,7 +220,7 @@ public class GlobalPoolingLayer extends NoParamLayer {
          * width) Default for CNN3D data: pooling dimensions 2,3,4 (depth, height and width)
          *
          */
-        private int[] poolingDimensions;
+        private long[] poolingDimensions;
 
         /**
          * P-norm constant. Only used if using {@link PoolingType#PNORM} for the pooling type
@@ -259,7 +259,7 @@ public class GlobalPoolingLayer extends NoParamLayer {
          *
          * @param poolingDimensions Pooling dimensions to use
          */
-        public Builder poolingDimensions(int... poolingDimensions) {
+        public Builder poolingDimensions(long... poolingDimensions) {
             this.setPoolingDimensions(poolingDimensions);
             return this;
         }
